+++ src/commands.rs
@@ -1,0 +1,4 @@
+pub mod ping;
+pub mod register_commands;
+pub mod register;
+
